+++ src/compiler/defblock.js
@@ -6,15 +6,14 @@ export default {
         const name = getStringLiteralValue( node.name );
         const placeholder = `${name}Block`;
         node.reference = placeholder;
-        figure.thisRef = true;
         figure.declare(
             sourceNode( `const ${placeholder} = document.createComment( '${name}' );` )
         );
 
         figure.addRenderActions(
             sourceNode( [
-                `            if ( _this.blocks[ '${name}' ] ) {\n`,
-                `                _this.blocks[ '${name}' ]( ${placeholder}, _this );\n`,
+                `            if ( this.blocks[ '${name}' ] ) {\n`,
+                `                this.blocks[ '${name}' ]( ${placeholder}, this );\n`,
                 '            }'
             ] )
         );